feature: Adiciona mais testes, script para rodar e tira warnings

+++ src/file_reader.cpp
@@ -7,7 +7,6 @@ void FileReader::readFile(string file_name){
   streampos fileSize;
 
   ifstream file_read(file_name, ios::binary);
-
   // get its size:
   file_read.seekg(0, ios::end);
   fileSize = file_read.tellg();
@@ -17,8 +16,8 @@ void FileReader::readFile(string file_name){
   vector<BYTE> fileData(fileSize);
   file_read.read((char *)&fileData[0], fileSize);
 
-  for (int i=0; i<fileData.size(); i++){
-    this->file.push_back(fileData[i]); 
+  for (unsigned int i=0; i<fileData.size(); i++){
+    this->file.push_back(fileData[i]);
   }
 
   file_read.close();
@@ -26,7 +25,7 @@ void FileReader::readFile(string file_name){
 
 vector<BYTE> FileReader::getBytes(int size){
   vector<BYTE> temp;
-  for(int i=this->position; i<this->position+size;i++){
+  for(unsigned int i=this->position; i<this->position+size;i++){
     temp.push_back(this->file[i]);
   }
   this->position += size;
@@ -35,7 +34,7 @@ vector<BYTE> FileReader::getBytes(int size){
 
 uint16_t FileReader::vectorTo2Bytes(vector<BYTE> bytes){
   uint16_t temp = 0;
-  for(int i=0; i< bytes.size();i++){
+  for(long unsigned int i=0; i< bytes.size();i++){
     BYTE masked_bytes = (bytes[i] & 0xff);
     int shift = ((bytes.size()-i-1)*8);
     temp |= masked_bytes << shift;
@@ -45,7 +44,7 @@ uint16_t FileReader::vectorTo2Bytes(vector<BYTE> bytes){
 
 uint32_t FileReader::vectorTo4Bytes(vector<BYTE> bytes){
   uint32_t temp = 0;
-  for(int i=0; i< bytes.size();i++){
+  for(long unsigned int i=0; i< bytes.size();i++){
     BYTE masked_bytes = (bytes[i] & 0xff);
     int shift = ((bytes.size()-i-1)*8);
     temp |= masked_bytes << shift;
